# pandora_console/api/v1/public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nueva API de Pandora FMS",
    "description": "Nueva Api de pandora FMS",
    "termsOfService": "https://example.com/terms/",
    "contact": {
      "name": "Nombre del contacto",
      "url": "https://www.example.com/support",
      "email": "contacto@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "PandoraFMS API Server"
    }
  ],
  "paths": {
...
    "/profile/{idProfile}": {
      "get": {
        "tags": ["Profiles"],
        "summary": "Show profile",
        "operationId": "e66704aa8cb757e23de04ce65c7bf028",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Profiles"],
        "summary": "Updates an profile",
        "operationId": "723fc236ab3b6069b7dcca9e317c953f",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyProfile"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseProfile"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Profiles"],
        "summary": "Deletes an profile object.",
        "operationId": "ca8ec2d296930aac798229f83e93bff8",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdProfile"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/profile/list": {
      "post": {
        "tags": ["Profiles"],
        "summary": "List profiles",
        "operationId": "875c11ea1ec723d55e2f582ed198db7a",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterPage"
          },
          {
            "$ref": "#/components/parameters/parameterSizePage"
          },
          {
            "$ref": "#/components/parameters/parameterSortField"
          },
          {
            "$ref": "#/components/parameters/parameterSortDirection"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyProfileFilter"
        },
        "responses": {
          "200": {
            "description": "List Incidence object",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "paginationData": {
                      "$ref": "#/components/schemas/paginationData"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Profile"
                      }
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/user/{idUser}": {
      "get": {
        "tags": ["Users"],
        "summary": "show users",
        "operationId": "aca8667503b4f47013622baeea8ecee3",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseUser"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Users"],
        "summary": "Updates an user",
        "operationId": "0a1f3fa99cb711d2ed1f47897d0e05f7",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/requestBodyUser"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/ResponseUser"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deletes an user object.",
        "operationId": "ac81c40c168ebda19372ceb213a4d422",
        "parameters": [
          {
            "$ref": "#/components/parameters/parameterIdUser"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/successfullyDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Profile": {
        "properties": {
          "idProfile": {
            "description": "Id Incidence Type",
            "type": "integer",
            "readOnly": true,
            "nullable": false
          },
          "name": {
            "description": "Name of the profile",
            "type": "string",
            "default": null,
            "nullable": false
          },
          "isAgentView": {
            "description": "Agent Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAgentEdit": {
            "description": "Agent Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAlertEdit": {
            "description": "Alert write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isUserManagement": {
            "description": "User Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isDbManagement": {
            "description": "Database Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAlertManagement": {
            "description": "Alert Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isPandoraManagement": {
            "description": "Pandora Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportView": {
            "description": "Report view",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportEdit": {
            "description": "Report Edit",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isReportManagement": {
            "description": "Report Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventView": {
            "description": "Event Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventEdit": {
            "description": "Event write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isEventManagement": {
            "description": "Event Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isAgentDisable": {
            "description": "Agent Disable",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapView": {
            "description": "Map Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapEdit": {
            "description": "Map Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isMapManagement": {
            "description": "Map Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleView": {
            "description": "Visual console Read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleEdit": {
            "description": "Visual console Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isVconsoleManagement": {
            "description": "Visual console Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigView": {
            "description": "Network config read",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigEdit": {
            "description": "Network config Write",
            "type": "boolean",
            "default": false,
            "nullable": true
          },
          "isNetworkConfigManagement": {
            "description": "Network config Management",
            "type": "boolean",
            "default": false,
            "nullable": true
          }
        },
        "type": "object"
      },
      "ProfileFilter": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Profile"
          },
          {
            "properties": {
              "idProfile": {
                "default": null,
                "readOnly": false
              },
              "freeSearch": {
                "description": "Find word in name field.",
                "type": "string",
                "default": null,
                "nullable": true
              }
            },
            "type": "object"
          }
        ]
      },
      "paginationData": {
        "description": "Info pagination data",
        "properties": {
          "totalPages": {
            "description": "Number of pages",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "sizePage": {
            "description": "Items per page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "totalRegisters": {
            "description": "Number of items",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "totalRegistersPage": {
            "description": "Number of items this page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          },
          "currentPage": {
            "description": "Number of current page",
            "type": "integer",
            "readOnly": true,
            "nullable": true
          }
        },
        "type": "object"
      },
...
    },
    "responses": {
      "ResponseProfile": {
        "description": "Profile object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad request",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "description": "Error",
              "properties": {
                "error": {
                  "type": "string",
                  "default": "Message error"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "successfullyDeleted": {
        "description": "Successfully deleted",
        "content": {
          "application/json": {
            "schema": {
              "properties": {
                "result": {
                  "type": "string",
                  "default": "Successfully deleted"
                }
              },
              "type": "object"
            }
          }
        }
      },
      "ResponseUser": {
        "description": "User object",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
    },
    "parameters": {
      "parameterIdProfile": {
        "name": "idProfile",
        "in": "path",
        "description": "Profile id",
        "required": true,
        "schema": {
          "type": "integer",
          "default": 1
        }
      },
      "parameterPage": {
        "name": "page",
        "in": "query",
        "description": "page",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        }
      },
      "parameterSizePage": {
        "name": "sizePage",
        "in": "query",
        "description": "Size page",
        "required": false,
        "schema": {
          "type": "integer",
          "default": 0
        }
      },
      "parameterSortField": {
        "name": "sortField",
        "in": "query",
        "description": "sort field",
        "required": false,
        "schema": {
          "type": "string",
          "default": ""
        }
      },
      "parameterSortDirection": {
        "name": "sortDirection",
        "in": "query",
        "description": "sort direction",
        "required": false,
        "schema": {
          "type": "string",
          "default": "",
          "enum": ["ascending", "descending"]
        }
      },
      "parameterIdUser": {
        "name": "idUser",
        "in": "path",
        "description": "User id",
        "required": true,
        "schema": {
          "type": "string",
          "default": "admin"
        }
      },
...
    },
    "requestBodies": {
      "requestBodyProfile": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Profile"
            }
          }
        }
      },
      "requestBodyProfileFilter": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProfileFilter"
            }
          }
        }
      },
      "requestBodyUser": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "in": "header",
        "bearerFormat": "pandoraBearer",
        "scheme": "bearer"
      }
    }
  },
...
}